requirements.txt
# requirements.txt
# Зависимости для курсового проекта "Мобильные устройства"

# Работа с базой данных PostgreSQL
psycopg2-binary==2.9.9

# GUI фреймворк
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# Работа с данными
pandas==2.1.4
numpy==1.26.2

# Для работы с CSV
openpyxl==3.1.2  # Опционально, для Excel файлов

# Логирование (встроено в Python, но можно расширить)
# colorlog==6.8.0  # Опционально, для цветного вывода логов